Added clay and it builds
