+++ DeckCAD.code-workspace
@@ -70,7 +70,10 @@
 			"__config": "cpp",
 			"iterator": "cpp",
 			"*.tcc": "cpp",
-			"*.m": "cpp"
+			"*.m": "cpp",
+			"clay_renderer_raylib.c": "cpp",
+			"clay-video-demo.c": "cpp",
+			"xlocale": "cpp"
 		}
 	},
 	"extensions": {
