+++ cmakelists.txt
@@ -5,7 +5,7 @@ set(CMAKE_C_STANDARD 11) # Requires C11 standard
 set(CMAKE_CXX_STANDARD 23) # Requires C++20 standard
 
 set(RAYLIB_DIR submodules/raylib/src)
-set(CLAY_DIR submodules/raylib/src)
+set(CLAY_DIR submodules/clay)
 
 add_executable(${PROJECT_NAME} src/main.cpp)
